Remove navigation drawer.
