+++ demo/src/main/java/com/google/android/fhir/demo/EditPatientFragment.kt
@@ -55,7 +55,6 @@ class EditPatientFragment : Fragment(R.layout.add_patient_fragment) {
       Toast.makeText(requireContext(), R.string.message_patient_updated, Toast.LENGTH_SHORT).show()
       NavHostFragment.findNavController(this).navigateUp()
     }
-    (activity as MainActivity).setDrawerEnabled(false)
 
     /** Use the provided cancel|submit buttons from the sdc library */
     childFragmentManager.setFragmentResultListener(
