+++ demo/src/main/java/com/google/android/fhir/demo/PatientDetailsFragment.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022-2023 Google LLC
+ * Copyright 2022-2024 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -82,7 +82,7 @@ class PatientDetailsFragment : Fragment() {
       }
     }
     patientDetailsViewModel.getPatientDetailData()
-    (activity as MainActivity).setDrawerEnabled(false)
+    //    (activity as MainActivity).setDrawerEnabled(false)
   }
 
   private fun onAddScreenerClick() {
