+++ demo/src/main/java/com/google/android/fhir/demo/PatientListFragment.kt
@@ -153,7 +153,7 @@ class PatientListFragment : Fragment() {
       addPatient.setColorFilter(Color.WHITE)
     }
     setHasOptionsMenu(true)
-    (activity as MainActivity).setDrawerEnabled(false)
+    //    (activity as MainActivity).setDrawerEnabled(false)
     launchAndRepeatStarted(
       { mainActivityViewModel.pollState.collect(::currentSyncJobStatus) },
       { mainActivityViewModel.pollPeriodicSyncJobStatus.collect(::periodicSyncJobStatus) },
